+++ examples/express.js
@@ -1,19 +1,19 @@
 import express from 'express';
 
-import { createError, errors, status } from '../src';
+import teapot from '../src';
 
 const app = express();
 
 app.get('/api', (req, res) => {
-    res.status(status.OK).json({
+    res.status(teapot.status.OK).json({
         api: 'running',
     });
 });
 
-app.get('/errors/404', (req, res, next) => next(new errors.NotFoundError('Nothing to see here!')));
-app.get('/errors/500', (req, res, next) => next(new errors.InternalServerError('Something broke!')));
-app.get('/custom/:code', (req, res, next) => next(createError(req.params.code)));
+app.get('/errors/404', (req, res, next) => next(new teapot.NotFoundError('Nothing to see here!')));
+app.get('/errors/500', (req, res, next) => next(new teapot.InternalServerError('Something broke!')));
+app.get('/custom/:code', (req, res, next) => next(teapot.error(req.params.code)));
 
 app.listen(3000, () => {
-    console.log('App listening on port 3000...');
+    console.log('App listening on port 3000...'); // eslint-disable-line no-console
 });